+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.5 - 2021-03-28
+## v0.5 - 2021-04-13
 
 - Major update and now with help from @Theile and @Isaac
 - Added additional `BLOCKING_PROCESS_ACTION` handlings